ソースファイルの分割と検索アルゴリズムの修正
