+++ tcusrs/makefile
@@ -1,0 +1,5 @@
+CC=GCC
+CFLAGS=-Wall
+
+tcusrs:main.o
+main.o:main.c check.c error_msg.c get_menu.c input_a_line.c ls_print.c search_result.c struct_init.c struct_load.c struct_print.c struct_search.c
